Redux書き換え完了
