+++ src/components/addTodo.js
@@ -1,16 +1,16 @@
-import React,{useState} from 'react'
-// ここからデータを送信するので、dispatch機能が必要
-import {dispatch} from 'react-redux'
+import React,{useSelector, useDispatch} from 'react-redux'
+import {useState} from 'react'
+
+// import { useDispatch } from "@reduxjs/toolkit";
+
+
+
 
 const addTodo = () => {
-  // 配列準備
   const [text, setText] = useState('')
-  // actionを運ぶ機能dispatch
   const dispatch = useDispatch()
 
   onClickAdd = (event) => setText(event.target.value)
-  
-  // keydownとtrimmedText
   onClickKeyDown = (event) => {
     const trimmedText = event.target.value.trim()
