--- src/store.js
@@ -1,6 +1,0 @@
-import {createStore} from 'redux'
-import rootReducer from './reducer'
-
-const store = createStore(rootReducer, dispatch)
-
-export default store
